+++ lapce-proxy/src/dispatch.rs
@@ -304,6 +304,7 @@ impl ProxyHandler for Dispatcher {
             } => {
                 let workspace = self.workspace.clone();
                 let proxy_rpc = self.proxy_rpc.clone();
+                // Perform the search on another thread to avoid blocking the proxy thread
                 thread::spawn(move || {
                     let result = if let Some(workspace) = workspace.as_ref() {
                         let mut matches = IndexMap::new();
@@ -325,10 +326,20 @@ impl ProxyHandler for Dispatcher {
                                                 let mymatch = matcher
                                                     .find(line.as_bytes())?
                                                     .unwrap();
+                                                // Shorten the line to avoid sending over absurdly long-lines
+                                                // (such as in minified javascript)
+                                                // Note that the start/end are column based, not absolute from the
+                                                // start of the file.
+                                                let display_range = mymatch
+                                                    .start()
+                                                    .saturating_sub(100)
+                                                    ..line
+                                                        .len()
+                                                        .min(mymatch.end() + 100);
                                                 line_matches.push((
                                                     lnum as usize,
                                                     (mymatch.start(), mymatch.end()),
-                                                    line.to_string(),
+                                                    line[display_range].to_string(),
                                                 ));
                                                 Ok(true)
                                             }),